+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Development
 
+- Expose the new `data_anonymized` field in the metadata schema as [blnal:dataAnonymized](https://berlin.github.io/lod-vocabulary/blnal/dataAnonmized).
 - Set env variables for database and Solr index in scripts for running unit tests locally.
 - Fix failing Github CI (install curl).